requirements.txt
torch
torchaudio
torchvision # apparently pytorch_lightning needs this (pyannote dependency)
transformers==4.36.1
accelerate
datasets
pyannote.audio
mutagen
nltk
jiwer
soundfile
mutagen
num2words
simplejson
whisperx
pyriksdagen
tqdm
rapidfuzz
ctc_segmentation